replace users-without-replication endpoint with replication-health/failed
