+++ api/src/services/replication/replication-health.js
@@ -3,6 +3,9 @@ const replicationLimitLog = require('./replication-limit-log');
 const replicationFailureLog = require('./replication-failure-log');
 
 const DAY_FORMAT = 'YYYY-MM-DD';
+// Default staleness window: a user who has not successfully replicated within this many days is
+// considered for the failed list when the caller doesn't supply an explicit `days` value.
+const DEFAULT_STALE_DAYS = 7;
 // "Any number of failures counts" — flag a user with at least one failure since their last log.
 const DEFAULT_MIN_FAILURES = 1;
 
@@ -13,18 +16,20 @@ const sumFailuresSince = (dailyFailures = {}, sinceDay) => {
 };
 
 /**
- * Lists users that appear unable to replicate: their replication limit log is older than a month
- * (so no successful replication has happened since) AND they have logged at least `minFailures`
- * replication failures since that last log. Failures are counted per user from the day of their
- * own last limit log onward, so the count reflects failures accrued since they last succeeded.
+ * Lists users that appear unable to replicate: their last successful replication predates `cutoff`
+ * AND they have logged at least `minFailures` replication failures since that last replication.
+ * Failures are counted per user from the day of their own last limit log onward.
  *
  * @param {object} [options]
+ * @param {number} [options.days=7] staleness window in days; users whose last successful replication
+ *   is older than this many days are considered. Defaults to DEFAULT_STALE_DAYS.
  * @param {number} [options.minFailures=1] minimum failures since the last log for a user to be listed
  * @returns {Promise<{ users: Array<{ user: string, last_replication_date: number,
  *   failures_since_last_replication: number }> }>}
  */
-const get = async ({ minFailures = DEFAULT_MIN_FAILURES } = {}) => {
-  const staleLogs = await replicationLimitLog.getStaleLogs();
+const getFailed = async ({ days = DEFAULT_STALE_DAYS, minFailures = DEFAULT_MIN_FAILURES } = {}) => {
+  const cutoff = moment().subtract(days, 'days').valueOf();
+  const staleLogs = await replicationLimitLog.getStaleLogs(cutoff);
   if (!staleLogs.length) {
     return { users: [] };
   }
@@ -47,6 +52,6 @@ const get = async ({ minFailures = DEFAULT_MIN_FAILURES } = {}) => {
 };
 
 module.exports = {
-  get,
+  getFailed,
   DEFAULT_MIN_FAILURES,
 };
